gh-925 remove fetch feature (replaced by explore)
