--- adapters/handlers/graphql/local/fetch/helpers_for_test.go
@@ -1,65 +1,0 @@
-//                           _       _
-// __      _____  __ ___   ___  __ _| |_ ___
-// \ \ /\ / / _ \/ _` \ \ / / |/ _` | __/ _ \
-//  \ V  V /  __/ (_| |\ V /| | (_| | ||  __/
-//   \_/\_/ \___|\__,_| \_/ |_|\__,_|\__\___|
-//
-//  Copyright © 2016 - 2019 Weaviate. All rights reserved.
-//  LICENSE: https://github.com/semi-technologies/weaviate/blob/develop/LICENSE.md
-//  DESIGN & CONCEPT: Bob van Luijt (@bobvanluijt)
-//  CONTACT: hello@semi.technology
-//
-
-package fetch
-
-import (
-	"context"
-
-	testhelper "github.com/semi-technologies/weaviate/adapters/handlers/graphql/test/helper"
-	"github.com/semi-technologies/weaviate/entities/models"
-	"github.com/semi-technologies/weaviate/usecases/traverser"
-	"github.com/stretchr/testify/mock"
-)
-
-type mockRequestsLog struct{}
-
-func (m *mockRequestsLog) Register(first string, second string) {
-
-}
-
-type mockResolver struct {
-	testhelper.MockResolver
-}
-
-func newMockResolver() *mockResolver {
-	field := Build()
-	mocker := &mockResolver{}
-	mockLog := &mockRequestsLog{}
-	mocker.RootFieldName = "Fetch"
-	mocker.RootField = field
-	mocker.RootObject = map[string]interface{}{
-		"Resolver":    Resolver(mocker),
-		"RequestsLog": mockLog,
-	}
-	return mocker
-}
-
-func (m *mockResolver) LocalFetchKindClass(ctx context.Context, principal *models.Principal,
-	params *traverser.FetchSearch) (interface{}, error) {
-	args := m.Called(params)
-	return args.Get(0), args.Error(1)
-}
-
-func (m *mockResolver) LocalFetchFuzzy(ctx context.Context, principal *models.Principal,
-	params traverser.FetchFuzzySearch) (interface{}, error) {
-	args := m.Called(params)
-	return args.Get(0), args.Error(1)
-}
-
-func newMockContextionary() *mockContextionary {
-	return &mockContextionary{}
-}
-
-type mockContextionary struct {
-	mock.Mock
-}
